Extracted method declaration
